refactor zkvm host script to derive from block
